fix: better player movement handling (#831)
This PR aims to standardize player position handling notably
teleportations and velocity changes.
Right now every module does its own thing and sends packets directly to
the client so it is impossible for the server to keep track of the
player's state.
By normalizing the way to change player position/velocity it could allow
us to implement better movement verification and better on_ground
calculation in the future.