fix: only allow one connection per account
Previously, if someone connected to the server twice with the same
account, the server would abort because two entities cannot have the
same name in flecs. In addition, having multiple players from the same
account would not make sense. This code disconnects the previous player
if a new player with the same name connects.

This also modifies IgnMap to no longer use DeferredMap. Deferred
insertions would allow for race conditions because two players with the
same name could be inserted into the IgnMap with neither insertion call
knowing that there's another player with the same name. The code also
cannot rely on flecs to determine if there's another player with the
same name using World::try_lookup because this would introduce a race
condition between the try_lookup call and setting the entity name.