fix(player_join): fix component race condition
There may be a period where the skin has been received through skins_rx
but the appropriate components (most importantly Position, as that must
be set through the event code and is out of the control of the hyperion
crate) have not yet been set, causing get to fail. Moving the
player_join_world call to its own system resolves this issue.