fix: malformed packet crash (#889)
When a client sends a malformed packet, the server will crash because it
used entity.destruct to disconnect players for malformed packets which
leads to an abort because, when the server receives PlayerDisconnect
from the proxy, the server will try to set PendingRemove on the
already-destructed entity.

This PR fixes these errors on player disconnects.