feat: add gui
This implements hyperion-gui and the associated gui commands.

The implementation of Gui::init has been removed for now; it would be
nontrivial to rewrite to Bevy, and it is also broken because it wouldn't
remove the ClickSlot handler when the gui is closed. It would be more
appropriate to fix these issues at a later time.