perf: avoid Bytes allocation on most packets
Bytes::from_owner has some performance cost because it needs to allocate
the metadata. However, the most frequently sent packets, such as the
position and look update packets, do not store Bytes in the packet and
are small enough to not be compressed and therefore do not require an
allocation to store the decompressed packet. This change avoids the
usage of Bytes::from_owner for these packets.