fix: make player teleport occur immediately
Player positions are updated immediately when TeleportEvent is sent
instead of later when the player confirms the teleportation. This allows
the new position to be immediately replicated to other clients, which is
also useful to allow the anticheat to continue working even if the client
does not acknowledge the teleportation quickly.

This also fixes a bug where a player's teleport is not synced
with other players because last_tick_position is set to the current
position every tick, meaning that the server will continue sending
position deltas relative to the previous location instead of the full
position to the teleport location.