refactor: add proxy channels and remove ordering (#937)
The server is able to create channels and send packets through those
channels. The proxy decides which clients are subscribed to those
channels and sends subscribe and unsubscribe packets appropriately.

Ordering is also removed. It added extra complexity which is no longer
needed since the order was changed to an incremental number. If ordering
was kept, the channels feature would be more complex because channel
creation and removal would also need to be ordered. With ordering fully
removed, systems are expected to use happens-before relationships to
determine the order of proxy messages.